backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
duckdb==1.2.1
pydantic==2.10.4
httpx==0.28.1
python-dotenv==1.0.1
